# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.1.0
  summary: Boilerplate HTTP API — auth, RBAC, users, roles, notifications, realtime
  description: |
    Starter HTTP API for the Next.js + Go boilerplate.

    **In scope:** authentication, global RBAC, platform users/roles, notifications,
    Web Push (VAPID), passkeys, and Centrifugo realtime tokens.

    **Out of scope:** tenants, workspaces, CRM, inbox, commerce, AI, tickets, reports.
    Those product modules were stripped. Live Go routes (`internal/modules/**/routes.go`)
    are the source of truth if a path below is missing from the router.

    **Roles:**
    - `super_admin` — system role; assigned via `make create-super-admin`
    - `organization_user` — system role for business owners/staff
    - Custom roles via `/v1/platform/roles`

    **Auth:** Bearer JWT access tokens + opaque refresh tokens.
    See `docs/auth.md` for the permission matrix and token claims.
    Error catalog: `docs/api/errors.md`.

    **Conventions:**
    - Success/error envelope: `{success,data|error,meta.request_id}` on all `/v1/*`.
    - Lists: `{items,total,limit,offset}` under `data` (`limit`/`offset`/`q`/`sort`; never `page`/`search`).
    - Probes `/healthz` and `/readyz` are **not** enveloped.
    - See `API_CONVENTIONS.md` and `docs/api/errors.md`.
  contact:
    name: App
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /v1/tenant/settings:
    get:
      tags: [Tenant]
      operationId: getTenantSettings
      summary: Get organization export letterhead
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Letterhead settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeTenantSettings"
    patch:
      tags: [Tenant]
      operationId: patchTenantSettings
      summary: Update organization export letterhead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTenantSettingsRequest"
      responses:
        "200":
          description: Updated letterhead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeTenantSettings"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ResponseMeta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string }
...
    TenantSettings:
      type: object
      required: [company_name, tagline, primary_color, address, city, district, phone, email, website, footer_text, paper_size]
      properties:
        company_name: { type: string }
        tagline: { type: string }
        primary_color: { type: string }
        address: { type: string }
        city: { type: string }
        district: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        footer_text: { type: string }
        paper_size: { type: string, enum: [A4, A3, Letter, Legal] }
        logo_url: { type: string, nullable: true }
    PatchTenantSettingsRequest:
      type: object
      properties:
        company_name: { type: string }
        tagline: { type: string }
        primary_color: { type: string }
        address: { type: string }
        city: { type: string }
        district: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        footer_text: { type: string }
        paper_size: { type: string, enum: [A4, A3, Letter, Legal] }
    EnvelopeTenantSettings:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: "#/components/schemas/TenantSettings" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }